[REFACTOR] AuthorRestController분리 및 javadoc 작성
