+++ src/main/java/store/ckin/front/author/controller/AuthorController.java
@@ -34,6 +34,13 @@ public class AuthorController {
     private final AuthorService authorService;
 
 
+    /**
+     * 모든 작가 정보를 조회하여 반환합니다.
+     *
+     * @param pageable 페이지 정보
+     * @param model    모델 객체
+     * @return 작가 목록 페이지 경로
+     */
     @GetMapping
     public String findAllAuthors(@PageableDefault Pageable pageable, Model model) {
         PageResponse<AuthorResponseDto> authors = authorService.getAuthors(pageable);
@@ -53,6 +60,12 @@ public class AuthorController {
         return "admin/author/index";
     }
 
+    /**
+     * 새로운 작가를 추가합니다.
+     *
+     * @param authorCreateRequestDto 작가 생성 요청 DTO
+     * @return 작가 목록 페이지로의 리다이렉트 경로
+     */
     @PostMapping
     public String addAuthor(@ModelAttribute AuthorCreateRequestDto authorCreateRequestDto) {
         authorService.createAuthor(authorCreateRequestDto);
@@ -60,6 +73,13 @@ public class AuthorController {
         return "redirect:/admin/authors";
     }
 
+    /**
+     * 주어진 ID의 작가 정보를 수정합니다.
+     *
+     * @param authorId 작가 ID
+     * @param authorModifyRequestDto 작가 수정 요청 DTO
+     * @return 작가 목록 페이지로의 리다이렉트 경로
+     */
     @PutMapping("/{authorId}")
     public String modifyAuthor(@PathVariable Long authorId,
                                @ModelAttribute AuthorModifyRequestDto authorModifyRequestDto) {
@@ -68,6 +88,14 @@ public class AuthorController {
         return "redirect:/admin/authors";
     }
 
+    /**
+     * 이름으로 작가를 검색하고 결과를 반환합니다.
+     *
+     * @param name     검색할 이름
+     * @param pageable 페이지 정보
+     * @param model    모델 객체
+     * @return 작가 목록 페이지 경로
+     */
     @GetMapping("/search")
     public String getAuthorsByName(@RequestParam String name, @PageableDefault Pageable pageable, Model model) {
         PageResponse<AuthorResponseDto> authors = authorService.getAuthorsByName(name, pageable);
@@ -89,22 +117,16 @@ public class AuthorController {
     }
 
 
-    //TODO: RestController로 분리
-    @GetMapping("/authorList")
-    @ResponseBody
-    public PageResponse<AuthorResponseDto> findAllAuthors(@PageableDefault Pageable pageable) {
-        return authorService.getAuthors(pageable);
-    }
-
-    @GetMapping("/authorList/search")
-    @ResponseBody
-    public PageResponse<AuthorResponseDto> getAuthorsByName(@RequestParam String name,
-                                                            @PageableDefault Pageable pageable) {
-        PageResponse<AuthorResponseDto> authors = authorService.getAuthorsByName(name, pageable);
-        return authors;
-    }
 
 
+
+
+    /**
+     * 주어진 ID의 작가를 삭제합니다.
+     *
+     * @param authorId 삭제할 작가 ID
+     * @return 작가 목록 페이지로의 리다이렉트 경로
+     */
     @DeleteMapping("/{authorId}")
     public String deleteAuthor(@PathVariable Long authorId) {
         authorService.deleteAuthor(authorId);
